splitting code to several files
