+++ main.py
@@ -12,7 +12,11 @@ elevation data, and field-of-view parameters.
 import argparse
 import logging
 import sys
-from pathlib import Path
+
+from src.InputProcessor import InputProcessor
+#from src.Elevation import ElevationService
+from src.VisibilityAnalyzer import VisibilityAnalyzer
+from src.Visualizer import Visualizer
 
 # Set up logging
 logging.basicConfig(
@@ -21,191 +25,6 @@ logging.basicConfig(
     handlers=[logging.StreamHandler(sys.stdout)]
 )
 logger = logging.getLogger('viewscape')
-
-
-class InputProcessor:
-    """Process user input parameters and prepare the map slice."""
-    
-    def __init__(self):
-        logger.info("Initializing InputProcessor")
-        
-    def process_coordinates(self, latitude, longitude, height, fov_degrees):
-        """
-        Process the input GPS coordinates and view parameters.
-        
-        Args:
-            latitude (float): Viewer's latitude in decimal degrees
-            longitude (float): Viewer's longitude in decimal degrees
-            height (float): Additional height above ground level in meters
-            fov_degrees (float): Field of view in degrees
-            
-        Returns:
-            dict: A dictionary containing the map slice parameters
-        """
-        logger.info(f"Processing coordinates: Lat {latitude}, Lon {longitude}, "
-                   f"Height {height}m, FOV {fov_degrees}°")
-        
-        # This is a placeholder - actual implementation would calculate
-        # the map slice based on the coordinates and FOV
-        map_slice = {
-            'center': (latitude, longitude),
-            'height': height,
-            'fov': fov_degrees,
-            'slice_bounds': {
-                'min_lat': latitude - 0.1,
-                'max_lat': latitude + 0.1,
-                'min_lon': longitude - 0.1,
-                'max_lon': longitude + 0.1
-            }
-        }
-        
-        return map_slice
-
-
-class VisibilityAnalyzer:
-    """Analyze what is visible from the given viewpoint."""
-    
-    def __init__(self):
-        logger.info("Initializing VisibilityAnalyzer")
-        
-    def fetch_elevation_data(self, map_slice):
-        """
-        Fetch elevation data for the given map slice.
-        
-        Args:
-            map_slice (dict): Map slice parameters
-            
-        Returns:
-            numpy.ndarray: 2D array of elevation data
-        """
-        logger.info("Fetching elevation data for map slice")
-        
-        # This is a placeholder - actual implementation would fetch
-        # elevation data from a DEM (Digital Elevation Model) source
-        # like SRTM data, LIDAR, or other elevation APIs
-        import numpy as np
-        
-        # Create a dummy elevation grid (in a real implementation, this would
-        # be actual elevation data from a GIS source)
-        grid_size = 100
-        elevation_data = np.zeros((grid_size, grid_size))
-        
-        # Add some sample terrain features (hills, valleys)
-        x = np.linspace(0, 10, grid_size)
-        y = np.linspace(0, 10, grid_size)
-        X, Y = np.meshgrid(x, y)
-        
-        # Create a sample terrain with some hills
-        elevation_data = 100 + 50 * np.sin(X) * np.cos(Y)
-        
-        return elevation_data
-    
-    def calculate_visibility(self, elevation_data, map_slice):
-        """
-        Calculate visibility from the viewpoint using the elevation data.
-        
-        Args:
-            elevation_data (numpy.ndarray): 2D array of elevation data
-            map_slice (dict): Map slice parameters
-            
-        Returns:
-            numpy.ndarray: 2D boolean array indicating visible (True) and non-visible (False) points
-        """
-        logger.info("Calculating visibility for viewpoint")
-        
-        # This is a placeholder - actual implementation would use
-        # line-of-sight algorithms to determine visibility
-        import numpy as np
-        
-        # Get map dimensions
-        height, width = elevation_data.shape
-        
-        # Create a visibility grid (initially all False)
-        visibility = np.zeros_like(elevation_data, dtype=bool)
-        
-        # In a real implementation, this would use ray casting or similar
-        # algorithms to determine what points are visible from the viewpoint
-        
-        # For demonstration, we'll just mark a circular area as visible
-        center_y, center_x = height // 2, width // 2
-        viewer_height = map_slice['height']
-        
-        for y in range(height):
-            for x in range(width):
-                # Simple distance-based visibility for demonstration
-                distance = np.sqrt((y - center_y)**2 + (x - center_x)**2)
-                
-                # Mark as visible if within a certain radius
-                # (this is a simplistic placeholder)
-                if distance < min(height, width) // 3:
-                    visibility[y, x] = True
-        
-        logger.info(f"Visibility calculated. {np.sum(visibility)} visible points out of {height*width}")
-        return visibility
-
-
-class Visualizer:
-    """Visualize the terrain and visibility data."""
-    
-    def __init__(self):
-        logger.info("Initializing Visualizer")
-    
-    def create_visibility_image(self, elevation_data, visibility, output_path=None):
-        """
-        Create a visualization of the terrain with visibility overlay.
-        
-        Args:
-            elevation_data (numpy.ndarray): 2D array of elevation data
-            visibility (numpy.ndarray): 2D boolean array of visibility
-            output_path (str, optional): Path to save the output image
-            
-        Returns:
-            str: Path to the saved image
-        """
-        logger.info("Creating visibility visualization")
-        
-        # This is a placeholder - actual implementation would create
-        # a more sophisticated visualization
-        import matplotlib.pyplot as plt
-        import numpy as np
-        from matplotlib.colors import LightSource
-        
-        fig, ax = plt.subplots(figsize=(10, 8))
-        
-        # Create a nice terrain visualization with hillshading
-        ls = LightSource(azdeg=315, altdeg=45)
-        rgb = ls.shade(elevation_data, cmap=plt.cm.terrain, vert_exag=0.3, blend_mode='soft')
-        
-        # Show the terrain
-        ax.imshow(rgb)
-        
-        # Create a semi-transparent overlay for visible areas
-        visibility_overlay = np.zeros((*visibility.shape, 4))  # RGBA
-        visibility_overlay[visibility, 3] = 0.5  # Alpha channel for visible areas
-        visibility_overlay[visibility, 0] = 0.0  # R
-        visibility_overlay[visibility, 1] = 0.8  # G
-        visibility_overlay[visibility, 2] = 0.0  # B
-        
-        ax.imshow(visibility_overlay)
-        
-        # Add title and labels
-        ax.set_title('Terrain Visibility Analysis')
-        
-        # Remove axes for cleaner look
-        ax.set_axis_off()
-        
-        # Save the figure if output_path is provided
-        if output_path:
-            output_path = Path(output_path)
-            output_path.parent.mkdir(parents=True, exist_ok=True)
-            plt.savefig(output_path, dpi=300, bbox_inches='tight')
-            logger.info(f"Visualization saved to {output_path}")
-            return str(output_path)
-        else:
-            # Display the figure
-            plt.tight_layout()
-            plt.show()
-            return None
 
 
 def parse_arguments():
