+++ src/InputProcessor.py
@@ -1,0 +1,53 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+
+"""
+InputProcessor module for ViewScape application.
+
+This module handles and validates user input including GPS coordinates,
+height parameters, and field-of-view angles. It calculates the appropriate
+map slice based on these inputs for further processing.
+"""
+
+import logging
+
+# Get the module logger
+logger = logging.getLogger('viewscape.input_processor')
+
+class InputProcessor:
+    """Process user input parameters and prepare the map slice."""
+    
+    def __init__(self):
+        logger.info("Initializing InputProcessor")
+        
+    def process_coordinates(self, latitude, longitude, height, fov_degrees):
+        """
+        Process the input GPS coordinates and view parameters.
+        
+        Args:
+            latitude (float): Viewer's latitude in decimal degrees
+            longitude (float): Viewer's longitude in decimal degrees
+            height (float): Additional height above ground level in meters
+            fov_degrees (float): Field of view in degrees
+            
+        Returns:
+            dict: A dictionary containing the map slice parameters
+        """
+        logger.info(f"Processing coordinates: Lat {latitude}, Lon {longitude}, "
+                   f"Height {height}m, FOV {fov_degrees}°")
+        
+        # This is a placeholder - actual implementation would calculate
+        # the map slice based on the coordinates and FOV
+        map_slice = {
+            'center': (latitude, longitude),
+            'height': height,
+            'fov': fov_degrees,
+            'slice_bounds': {
+                'min_lat': latitude - 0.1,
+                'max_lat': latitude + 0.1,
+                'min_lon': longitude - 0.1,
+                'max_lon': longitude + 0.1
+            }
+        }
+        
+        return map_slice
