+++ src/VisibilityAnalyzer.py
@@ -1,0 +1,95 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+
+"""
+VisibilityAnalyzer module for ViewScape application.
+
+This module implements line-of-sight algorithms to determine which points
+in a terrain model are visible from a specified viewpoint.
+"""
+
+import logging
+
+# Get the module logger
+logger = logging.getLogger('viewscape.visibility_analyzer')
+
+class VisibilityAnalyzer:
+    """Analyze what is visible from the given viewpoint."""
+    
+    def __init__(self):
+        logger.info("Initializing VisibilityAnalyzer")
+        
+    def fetch_elevation_data(self, map_slice):
+        """
+        Fetch elevation data for the given map slice.
+        
+        Args:
+            map_slice (dict): Map slice parameters
+            
+        Returns:
+            numpy.ndarray: 2D array of elevation data
+        """
+        logger.info("Fetching elevation data for map slice")
+        
+        # This is a placeholder - actual implementation would fetch
+        # elevation data from a DEM (Digital Elevation Model) source
+        # like SRTM data, LIDAR, or other elevation APIs
+        import numpy as np
+        
+        # Create a dummy elevation grid (in a real implementation, this would
+        # be actual elevation data from a GIS source)
+        grid_size = 100
+        elevation_data = np.zeros((grid_size, grid_size))
+        
+        # Add some sample terrain features (hills, valleys)
+        x = np.linspace(0, 10, grid_size)
+        y = np.linspace(0, 10, grid_size)
+        X, Y = np.meshgrid(x, y)
+        
+        # Create a sample terrain with some hills
+        elevation_data = 100 + 50 * np.sin(X) * np.cos(Y)
+        
+        return elevation_data
+    
+    def calculate_visibility(self, elevation_data, map_slice):
+        """
+        Calculate visibility from the viewpoint using the elevation data.
+        
+        Args:
+            elevation_data (numpy.ndarray): 2D array of elevation data
+            map_slice (dict): Map slice parameters
+            
+        Returns:
+            numpy.ndarray: 2D boolean array indicating visible (True) and non-visible (False) points
+        """
+        logger.info("Calculating visibility for viewpoint")
+        
+        # This is a placeholder - actual implementation would use
+        # line-of-sight algorithms to determine visibility
+        import numpy as np
+        
+        # Get map dimensions
+        height, width = elevation_data.shape
+        
+        # Create a visibility grid (initially all False)
+        visibility = np.zeros_like(elevation_data, dtype=bool)
+        
+        # In a real implementation, this would use ray casting or similar
+        # algorithms to determine what points are visible from the viewpoint
+        
+        # For demonstration, we'll just mark a circular area as visible
+        center_y, center_x = height // 2, width // 2
+        viewer_height = map_slice['height']
+        
+        for y in range(height):
+            for x in range(width):
+                # Simple distance-based visibility for demonstration
+                distance = np.sqrt((y - center_y)**2 + (x - center_x)**2)
+                
+                # Mark as visible if within a certain radius
+                # (this is a simplistic placeholder)
+                if distance < min(height, width) // 3:
+                    visibility[y, x] = True
+        
+        logger.info(f"Visibility calculated. {np.sum(visibility)} visible points out of {height*width}")
+        return visibility
