+++ src/Visualizer.py
@@ -1,0 +1,79 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+
+"""
+Visualizer module for ViewScape application.
+
+This module creates visual representations of terrain data and visibility
+analysis results, with options for different visualization styles and
+output formats.
+"""
+
+import logging
+from pathlib import Path
+
+# Get the module logger
+logger = logging.getLogger('viewscape.visualizer')
+
+class Visualizer:
+    """Visualize the terrain and visibility data."""
+    
+    def __init__(self):
+        logger.info("Initializing Visualizer")
+    
+    def create_visibility_image(self, elevation_data, visibility, output_path=None):
+        """
+        Create a visualization of the terrain with visibility overlay.
+        
+        Args:
+            elevation_data (numpy.ndarray): 2D array of elevation data
+            visibility (numpy.ndarray): 2D boolean array of visibility
+            output_path (str, optional): Path to save the output image
+            
+        Returns:
+            str: Path to the saved image
+        """
+        logger.info("Creating visibility visualization")
+        
+        # This is a placeholder - actual implementation would create
+        # a more sophisticated visualization
+        import matplotlib.pyplot as plt
+        import numpy as np
+        from matplotlib.colors import LightSource
+        
+        fig, ax = plt.subplots(figsize=(10, 8))
+        
+        # Create a nice terrain visualization with hillshading
+        ls = LightSource(azdeg=315, altdeg=45)
+        rgb = ls.shade(elevation_data, cmap=plt.cm.terrain, vert_exag=0.3, blend_mode='soft')
+        
+        # Show the terrain
+        ax.imshow(rgb)
+        
+        # Create a semi-transparent overlay for visible areas
+        visibility_overlay = np.zeros((*visibility.shape, 4))  # RGBA
+        visibility_overlay[visibility, 3] = 0.5  # Alpha channel for visible areas
+        visibility_overlay[visibility, 0] = 0.0  # R
+        visibility_overlay[visibility, 1] = 0.8  # G
+        visibility_overlay[visibility, 2] = 0.0  # B
+        
+        ax.imshow(visibility_overlay)
+        
+        # Add title and labels
+        ax.set_title('Terrain Visibility Analysis')
+        
+        # Remove axes for cleaner look
+        ax.set_axis_off()
+        
+        # Save the figure if output_path is provided
+        if output_path:
+            output_path = Path(output_path)
+            output_path.parent.mkdir(parents=True, exist_ok=True)
+            plt.savefig(output_path, dpi=300, bbox_inches='tight')
+            logger.info(f"Visualization saved to {output_path}")
+            return str(output_path)
+        else:
+            # Display the figure
+            plt.tight_layout()
+            plt.show()
+            return None
